+++ apps/web/src/components/Comment/Feed.tsx
@@ -14,8 +14,8 @@ import { OptmisticPublicationType } from '@hey/types/enums';
 import { Card, EmptyState, ErrorMessage } from '@hey/ui';
 import { type FC } from 'react';
 import { useInView } from 'react-cool-inview';
+import { useImpressionsStore } from 'src/store/non-persisted/useImpressionsStore';
 import { useTransactionStore } from 'src/store/persisted/useTransactionStore';
-import { useImpressionsStore } from 'src/store/useImpressionsStore';
 
 interface FeedProps {
   publication: AnyPublication;
