+++ apps/web/src/components/Common/Providers/FeatureFlagsProvider.tsx
@@ -4,7 +4,7 @@ import getCurrentSession from '@lib/getCurrentSession';
 import { useQuery } from '@tanstack/react-query';
 import axios from 'axios';
 import { type FC } from 'react';
-import { useFeatureFlagsStore } from 'src/store/useFeatureFlagsStore';
+import { useFeatureFlagsStore } from 'src/store/non-persisted/useFeatureFlagsStore';
 import { isAddress } from 'viem';
 
 const FeatureFlagsProvider: FC = () => {
