+++ apps/web/src/components/Composer/Actions/CollectSettings/SplitConfig.tsx
@@ -12,8 +12,8 @@ import { OpenActionModuleType } from '@hey/lens';
 import splitNumber from '@hey/lib/splitNumber';
 import { Button, Input } from '@hey/ui';
 import { type FC } from 'react';
-import useProfileStore from 'src/store/persisted/useProfileStore';
 import { useCollectModuleStore } from 'src/store/non-persisted/useCollectModuleStore';
+import useProfileStore from 'src/store/persisted/useProfileStore';
 import { isAddress } from 'viem';
 
 interface SplitConfigProps {
