+++ apps/web/src/components/Explore/Feed.tsx
@@ -15,7 +15,7 @@ import {
 import { Card, EmptyState, ErrorMessage } from '@hey/ui';
 import { type FC } from 'react';
 import { useInView } from 'react-cool-inview';
-import { useImpressionsStore } from 'src/store/useImpressionsStore';
+import { useImpressionsStore } from 'src/store/non-persisted/useImpressionsStore';
 
 interface FeedProps {
   focus?: PublicationMetadataMainFocusType;
