+++ apps/web/src/components/Group/Feed.tsx
@@ -12,7 +12,7 @@ import type { Group } from '@hey/types/hey';
 import { Card, EmptyState, ErrorMessage } from '@hey/ui';
 import { type FC } from 'react';
 import { useInView } from 'react-cool-inview';
-import { useImpressionsStore } from 'src/store/useImpressionsStore';
+import { useImpressionsStore } from 'src/store/non-persisted/useImpressionsStore';
 
 interface FeedProps {
   group: Group;
