+++ apps/web/src/components/Home/Highlights.tsx
@@ -8,10 +8,10 @@ import { OptmisticPublicationType } from '@hey/types/enums';
 import { Card, EmptyState, ErrorMessage } from '@hey/ui';
 import { type FC } from 'react';
 import { useInView } from 'react-cool-inview';
+import { useImpressionsStore } from 'src/store/non-persisted/useImpressionsStore';
 import { useTimelineStore } from 'src/store/non-persisted/useTimelineStore';
 import useProfileStore from 'src/store/persisted/useProfileStore';
 import { useTransactionStore } from 'src/store/persisted/useTransactionStore';
-import { useImpressionsStore } from 'src/store/useImpressionsStore';
 
 const Highlights: FC = () => {
   const currentProfile = useProfileStore((state) => state.currentProfile);
