+++ apps/web/src/components/Profile/Feed.tsx
@@ -13,7 +13,7 @@ import { Card, EmptyState, ErrorMessage } from '@hey/ui';
 import { type FC } from 'react';
 import { useInView } from 'react-cool-inview';
 import { ProfileFeedType } from 'src/enums';
-import { useImpressionsStore } from 'src/store/useImpressionsStore';
+import { useImpressionsStore } from 'src/store/non-persisted/useImpressionsStore';
 import { useProfileFeedStore } from 'src/store/useProfileFeedStore';
 
 interface FeedProps {
