+++ apps/web/src/components/Profile/Menu/Report.tsx
@@ -2,7 +2,7 @@ import { FlagIcon } from '@heroicons/react/24/outline';
 import type { Profile } from '@hey/lens';
 import cn from '@hey/ui/cn';
 import { type FC } from 'react';
-import { useGlobalModalStateStore } from 'src/store/useGlobalModalStateStore';
+import { useGlobalModalStateStore } from 'src/store/non-persisted/useGlobalModalStateStore';
 
 interface ReportProfileProps {
   profile: Profile;
