+++ apps/web/src/components/Publication/Actions/Share/Quote.tsx
@@ -4,7 +4,7 @@ import { type AnyPublication, TriStateValue } from '@hey/lens';
 import { isMirrorPublication } from '@hey/lib/publicationHelpers';
 import cn from '@hey/ui/cn';
 import { type FC } from 'react';
-import { useGlobalModalStateStore } from 'src/store/useGlobalModalStateStore';
+import { useGlobalModalStateStore } from 'src/store/non-persisted/useGlobalModalStateStore';
 import { usePublicationStore } from 'src/store/usePublicationStore';
 
 interface QuoteProps {
