+++ apps/web/src/components/Publication/Actions/index.tsx
@@ -4,9 +4,9 @@ import isOpenActionAllowed from '@hey/lib/isOpenActionAllowed';
 import { isMirrorPublication } from '@hey/lib/publicationHelpers';
 import stopEventPropagation from '@hey/lib/stopEventPropagation';
 import { type FC, memo } from 'react';
+import { useFeatureFlagsStore } from 'src/store/non-persisted/useFeatureFlagsStore';
+import { useImpressionsStore } from 'src/store/non-persisted/useImpressionsStore';
 import useProfileStore from 'src/store/persisted/useProfileStore';
-import { useFeatureFlagsStore } from 'src/store/useFeatureFlagsStore';
-import { useImpressionsStore } from 'src/store/useImpressionsStore';
 
 import OpenAction from '../LensOpenActions';
 import Comment from './Comment';
