+++ apps/web/src/components/Shared/Login/WrongNetwork.tsx
@@ -1,6 +1,6 @@
 import type { FC } from 'react';
 import { CHAIN_ID } from 'src/constants';
-import { useGlobalModalStateStore } from 'src/store/useGlobalModalStateStore';
+import { useGlobalModalStateStore } from 'src/store/non-persisted/useGlobalModalStateStore';
 
 import SwitchNetwork from '../SwitchNetwork';
