+++ apps/web/src/components/Shared/Navbar/NavItems/GardenerMode.tsx
@@ -8,7 +8,7 @@ import { Leafwatch } from '@lib/leafwatch';
 import axios from 'axios';
 import { type FC } from 'react';
 import { toast } from 'react-hot-toast';
-import { useFeatureFlagsStore } from 'src/store/useFeatureFlagsStore';
+import { useFeatureFlagsStore } from 'src/store/non-persisted/useFeatureFlagsStore';
 
 interface ModModeProps {
   className?: string;
