+++ apps/web/src/components/Shared/Navbar/NavItems/Invites.tsx
@@ -3,7 +3,7 @@ import { INVITE } from '@hey/data/tracking';
 import cn from '@hey/ui/cn';
 import { Leafwatch } from '@lib/leafwatch';
 import { type FC } from 'react';
-import { useGlobalModalStateStore } from 'src/store/useGlobalModalStateStore';
+import { useGlobalModalStateStore } from 'src/store/non-persisted/useGlobalModalStateStore';
 
 interface InvitesProps {
   className?: string;
