+++ apps/web/src/store/non-persisted/useGlobalModalStateStore.ts
